Rename to `now-layer-node`
