+++ .eslintignore
@@ -6,7 +6,7 @@
 /packages/now-build-utils/src/*.js
 /packages/now-build-utils/src/fs/*.js
 /packages/now-node/dist/*
-/packages/now-node-runtime/dist/*
+/packages/now-layer-node/dist/*
 /packages/now-next/dist/*
 /packages/now-node-bridge/*
 /packages/now-python/dist/*
