+++ packages/now-layer-node/types/promisepipe.ts
@@ -1,0 +1,4 @@
+declare module 'promisepipe' {
+	import { Stream } from 'stream';
+	export default function pipe(...args: Stream[]): Promise<void>;
+}
