+++ packages/now-layer-node/types/stream-to-promise.ts
@@ -1,0 +1,6 @@
+declare module 'stream-to-promise' {
+	import { Stream } from 'stream';
+	export default function streamToPromise(
+		stream: NodeJS.ReadableStream
+	): Promise<string>;
+}
