--- packages/now-node-runtime/types/promisepipe.ts
@@ -1,4 +1,0 @@
-declare module 'promisepipe' {
-  import { Stream } from 'stream';
-  export default function pipe(...args: Stream[]): Promise<void>;
-}
