--- packages/now-node-runtime/types/stream-to-promise.ts
@@ -1,6 +1,0 @@
-declare module 'stream-to-promise' {
-  import { Stream } from 'stream';
-  export default function streamToPromise(
-    stream: NodeJS.ReadableStream
-  ): Promise<string>;
-}
